+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
   **NOTE:** This raises the minimum supported Rust version of this crate from Rust 1.37 to Rust 1.46.
 
+- Detect misuse of `AssertUnmoved::get_mut`.
+
 ## [0.1.3] - 2021-04-06
 
 - [Apply `doc(cfg(...))` on feature gated APIs.](https://github.com/taiki-e/assert-unmoved/pull/3)